pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study-with-ai"
version = "0.1.0"
description = "Desktop MVP for AI-assisted study workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pdfminer-six",
  "pypdf",
  "PySide6>=6.8,<7",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.8,<4",
  "pyright>=1.1.390,<2",
  "pytest>=8.3,<9",
  "ruff>=0.8,<1",
]

[project.scripts]
praktikum-app = "praktikum_app.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
praktikum_app = ["assets/theme/*.qss", "assets/fonts/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "C4", "E", "F", "I", "N", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
include = ["src"]
exclude = [
  ".venv",
  ".pytest_cache",
  ".ruff_cache",
  "__pycache__",
]
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = "none"
